Fix UI layout for lab submission

+++ flatMachine/RoundedButton.swift
@@ -36,7 +36,7 @@ extension UIButton {
         }
         
         set {
-            cornerRadius = newValue
+            cornerRadius = CGFloat(newValue)
         }
     }
